EES-7041 - Changes to UserReleaseRoleQueriesTests
